storybook: Refactor and rename apollo decorators to be more flexible and use apolloStoryDecroator with a different api
