+++ packages/admin/admin-stories/src/admin/form/ScrollToErrorField.tsx
@@ -3,7 +3,7 @@ import { Button, Typography } from "@mui/material";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../apollo-rest-story.decorator";
 
 interface FormValues {
     foo: string;
@@ -56,5 +56,5 @@ function Story() {
 }
 
 storiesOf("@comet/admin/form", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .add("Scroll To Error Field", () => <Story />);
