+++ packages/admin/admin-stories/src/admin/form/Tabs.tsx
@@ -3,7 +3,7 @@ import { Button, Card, CardContent } from "@mui/material";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../apollo-rest-story.decorator";
 
 function Story() {
     const [showExample3, setShowExample3] = React.useState(false);
@@ -50,5 +50,5 @@ function Story() {
 }
 
 storiesOf("@comet/admin/form", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .add("Tabs", () => <Story />);
