+++ packages/admin/admin-stories/src/admin/table/TableBesidesFormSelectionHooks.tsx
@@ -16,7 +16,7 @@ import { storiesOf } from "@storybook/react";
 import * as React from "react";
 import { Redirect, Route, Switch, useLocation } from "react-router";
 
-import { apolloStoryDecorator } from "../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../apollo-rest-story.decorator";
 import { storyRouterDecorator } from "../../story-router.decorator";
 
 const gqlRest = gql;
@@ -141,6 +141,6 @@ function App() {
 }
 
 storiesOf("@comet/admin/table", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .addDecorator(storyRouterDecorator())
     .add("Besides Form Selection Hooks", () => <App />);
