+++ packages/admin/admin-stories/src/admin/table/TablePaging.tsx
@@ -3,7 +3,7 @@ import { createRestPagingActions, MainContent, Table, TableQuery, useTableQuery,
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../apollo-rest-story.decorator";
 
 const gqlRest = gql;
 
@@ -80,7 +80,7 @@ function Story() {
 
 storiesOf("@comet/admin/table", module)
     .addDecorator(
-        apolloStoryDecorator({
+        apolloRestStoryDecorator({
             uri: "https://swapi.co/api/",
         }),
     )
