+++ packages/admin/admin-stories/src/admin/table/TableResetFilter.tsx
@@ -6,7 +6,7 @@ import { storiesOf } from "@storybook/react";
 import * as qs from "qs";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../apollo-rest-story.decorator";
 
 const gqlRest = gql;
 
@@ -119,5 +119,5 @@ function Story() {
 }
 
 storiesOf("@comet/admin/table", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .add("Reset Filter", () => <Story />);
