+++ packages/admin/admin-stories/src/admin/table/TableSort.tsx
@@ -3,7 +3,7 @@ import { SortDirection, Table, TableQuery, useTableQuery, useTableQuerySort } fr
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../apollo-rest-story.decorator";
 
 const gqlRest = gql;
 
@@ -86,5 +86,5 @@ function Story() {
 }
 
 storiesOf("@comet/admin/table", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .add("Sort", () => <Story />);
