+++ packages/admin/admin-stories/src/admin/table/TableStackEditDialog.tsx
@@ -20,7 +20,7 @@ import { Button, IconButton, Typography } from "@mui/material";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../apollo-rest-story.decorator";
 import { storyRouterDecorator } from "../../story-router.decorator";
 
 interface IExampleRow {
@@ -140,5 +140,5 @@ function Story() {
 
 storiesOf("@comet/admin/table", module)
     .addDecorator(storyRouterDecorator())
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .add("Stack+EditDialog", () => <Story />);
