+++ packages/admin/admin-stories/src/docs/components/ErrorHandling/ErrorDialog/stories/AutomaticGraphqlErrorDialog.stories.tsx
@@ -4,7 +4,7 @@ import { storiesOf } from "@storybook/react";
 import gql from "graphql-tag";
 import * as React from "react";
 
-import { apolloSwapiStoryDecorator } from "../../../../../apollo-swapi-story.decorator";
+import { apolloSwapiStoryDecorator } from "../../../../../apollo-story.decorator";
 import { errorDialogStoryProviderDecorator } from "../error-dialog-provider.decorator";
 
 storiesOf("stories/components/Error Handling/Error Dialog/Automatic Graphql Error Dialog", module)
