+++ packages/admin/admin-stories/src/docs/components/Table/stories/SelectionTable.stories.tsx
@@ -4,7 +4,7 @@ import { Grid } from "@mui/material";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../../../apollo-rest-story.decorator";
 import { storyRouterDecorator } from "../../../../story-router.decorator";
 
 const gqlRest = gql;
@@ -53,7 +53,7 @@ const ExampleForm = (props: ExampleFormProps) => {
 };
 
 storiesOf("stories/components/Table/Selection Table", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .addDecorator(storyRouterDecorator())
     .add("Selection Table", () => {
         // step 1
