+++ packages/admin/admin-stories/src/docs/form/stories/FinalForm.stories.tsx
@@ -3,10 +3,10 @@ import { useFormApiRef } from "@comet/admin/lib/FinalForm";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../../apollo-rest-story.decorator";
 
 storiesOf("stories/form/FinalForm", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .add("Basic FinalForm", () => {
         return (
             <FinalForm
