+++ packages/admin/admin-stories/src/docs/form/stories/Validation.stories.tsx
@@ -3,10 +3,10 @@ import { storiesOf } from "@storybook/react";
 import { FORM_ERROR } from "final-form";
 import * as React from "react";
 
-import { apolloStoryDecorator } from "../../../apollo-story.decorator";
+import { apolloRestStoryDecorator } from "../../../apollo-rest-story.decorator";
 
 storiesOf("stories/Form/Validation", module)
-    .addDecorator(apolloStoryDecorator())
+    .addDecorator(apolloRestStoryDecorator())
     .add("Record-level validation", () => {
         interface FormValues {
             name: string;
